requirements.txt
streamlit
psycopg2-binary
bcrypt
streamlit-cookies-manager
pandas
plotly
python-dateutil
reportlab
